ref: organization code

+++ AssetsManager/Views/Models/Dialogs/Controls/WadResultsTreeModel.cs
@@ -126,7 +126,7 @@ namespace AssetsManager.Views.Models.Dialogs.Controls
         public int Count { get; set; }
         public double Percentage { get; set; }
         public long TotalSizeChange { get; set; }
-        public string SizeChangeText => (TotalSizeChange >= 0 ? "+" : "") + FormatUtils.FormatSize((ulong)Math.Abs(TotalSizeChange));
+        public string SizeChangeText => (TotalSizeChange >= 0 ? "+" : "") + FormatUtils.FormatSize(Math.Abs(TotalSizeChange));
     }
 
     public class TopImpactFile
@@ -137,7 +137,7 @@ namespace AssetsManager.Views.Models.Dialogs.Controls
         public ulong OldSize { get; set; }
         public ulong NewSize { get; set; }
         public long SizeDiff { get; set; }
-        public string SizeDiffText => (SizeDiff >= 0 ? "+" : "") + FormatUtils.FormatSize((ulong)Math.Abs(SizeDiff));
+        public string SizeDiffText => (SizeDiff >= 0 ? "+" : "") + FormatUtils.FormatSize(Math.Abs(SizeDiff));
     }
 
     public class AffectedArea
